添加WebSocket 和SSL支持，并且增加了WorkerManager，完善了日志模块

+++ sylar/stream.cpp
@@ -4,7 +4,7 @@
  * @Author: zsj
  * @Date: 2020-06-16 16:09:49
  * @LastEditors: zsj
- * @LastEditTime: 2020-06-16 16:14:17
+ * @LastEditTime: 2020-06-28 21:42:44
  */ 
 #include"stream.h"
 
@@ -13,9 +13,9 @@ namespace sylar
 
 int Stream::readFixSize(void * buffer,size_t length){
     size_t offset = 0;
-    size_t left = length;
+    int64_t left = length;
     while(left > 0){
-        size_t len = read((char*)buffer + offset,left);
+        int64_t len = read((char*)buffer + offset,left);
         if(len <= 0){
             return len;
         }
@@ -25,9 +25,9 @@ int Stream::readFixSize(void * buffer,size_t length){
     return length;
 }
 int Stream::readFixSize(ByteArray::ptr ba,size_t length){
-    size_t left = length;
+    int64_t left = length;
     while(left > 0){
-        size_t len = read(ba,left);
+        int64_t len = read(ba,left);
         if(len <= 0){
             return len;
         }
@@ -38,9 +38,9 @@ int Stream::readFixSize(ByteArray::ptr ba,size_t length){
 
 int Stream::writeFixSize(const void * buffer,size_t length){
     size_t offset = 0;
-    size_t left = length;
+    int64_t left = length;
     while(left > 0){
-        size_t len = write((const char*)buffer + offset,left);
+        int64_t len = write((const char*)buffer + offset,left);
         if(len <= 0){
             return len;
         }
@@ -50,9 +50,9 @@ int Stream::writeFixSize(const void * buffer,size_t length){
     return length;
 }
 int Stream::writeFixSize(ByteArray::ptr ba,size_t length){
-    size_t left = length;
+    int64_t left = length;
     while(left > 0){
-        size_t len = write(ba,left);
+        int64_t len = write(ba,left);
         if(len <= 0){
             return len;
         }
